requirements.txt
# Dependensi untuk Sistem Klasifikasi Kematangan Tomat
# Riset Jurnal: Otomasi Grading Kematangan Tomat Berbasis Deep Learning

# Deep Learning Framework
tensorflow>=2.10.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# Property-Based Testing
hypothesis>=6.0.0
